requirements.txt
mindspore>=2.6.0
numpy==1.26.4
pandas==2.3.3
scipy==1.15.3
matplotlib==3.10.8
pvlib==0.13.1
pytest==9.0.2
httpx
h5py==3.15.1
pyyaml>=6.0
passlib[bcrypt]
bcrypt<5.0.0
pydantic
pydantic-settings
sqlalchemy
python-dotenv
email-validator
importlib-metadata
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
